requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
mediapipe>=0.10.0
flask>=2.3.0
numpy>=1.24.0
pillow>=10.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0

